+++ pyhearts/feature/variability.py
@@ -138,3 +138,4 @@ def compute_beat_to_beat_variability(
     
     return variability_metrics
 
+